+++ packages/sdk/src/core/base.ts
@@ -96,4 +96,8 @@ export class BaseContract {
       };
     }
   }
+
+  setAppId(appId: string) {
+    this.appId = appId;
+  }
 }